requirements.txt
django>=3.0.0
psycopg2-binary
Faker>=3.0.0
pillow>=6.2.1
django-filter>=2.2.0
